試驗照功能分開。
