+++ 試驗/教典/test其他來源試驗.py
@@ -1,0 +1,9 @@
+from 臺灣言語工具.基本物件.字 import 字
+from 用字 import 教典
+from unittest.case import TestCase
+
+
+class 用字試驗(TestCase):
+
+    def test_牽水tsn̄g(self):
+        self.assertTrue(教典.有這个字無(字('𰹬', 'tsn̄g')))
